fix: createContext errors on profile page
moved charts into client component
